Bring back WinToastImpl with an improved version
Bring back WinToastImpl removed in
8876e46d52662baefecf66a2ff2f610372f2a230 but with improved version like
making everything static instead of an `instance` static method.

The pimpl class itself is returned back again as other classes will need
to become `friends` and access the private members like the AUMI that
was set.
