+++ CMakeLists.txt
@@ -5,6 +5,7 @@ set(CMAKE_CXX_STANDARD 17)
 
 add_library(WinToast STATIC
         src/wintoast.cpp
+        src/wintoast_impl.cpp
         src/win_toast_arguments.cpp
         src/win_toast_template.cpp)
 target_include_directories(WinToast PRIVATE
